+++ R/method-fortify.R
@@ -1,8 +1,123 @@
+
+
+##' convert compareClusterResult to a data.frame that ready for plot
+##'
+##'
+##' @rdname fortify
+##' @title fortify
+##' @param includeAll logical
+##' @return data.frame
+##' @importFrom ggplot2 fortify
+##' @importFrom plyr ddply
+##' @importFrom plyr mdply
+##' @importFrom plyr .
+##' @method fortify compareClusterResult
+##' @export
+##' @author Guangchuang Yu
+fortify.compareClusterResult <- function(model, data, showCategory=5, by="geneRatio",
+                                         split=NULL, includeAll=TRUE) {
+    clProf.df <- as.data.frame(model)
+    .split <- split
+
+    ## get top 5 (default) categories of each gene cluster.
+    if (is.null(showCategory)) {
+        result <- clProf.df
+    } else {
+        Cluster <- NULL # to satisfy codetools
+
+        topN <- function(res, showCategory) {
+            ddply(.data = res,
+                  .variables = .(Cluster),
+                  .fun = function(df, N) {
+                      if (length(df$Count) > N) {
+                          if (any(colnames(df) == "pvalue")) {
+                              idx <- order(df$pvalue, decreasing=FALSE)[1:N]
+                          } else {
+                              ## for groupGO
+                              idx <- order(df$Count, decreasing=T)[1:N]
+                          }
+                          return(df[idx,])
+                      } else {
+                          return(df)
+                      }
+                  },
+                  N=showCategory
+                  )
+
+        }
+
+        if (!is.null(.split) && .split %in% colnames(clProf.df)) {
+            lres <- split(clProf.df, as.character(clProf.df[, .split]))
+            lres <- lapply(lres, topN, showCategory = showCategory)
+            result <- do.call('rbind', lres)
+        } else {
+            result <- topN(clProf.df, showCategory)
+        }
+
+    }
+
+    ID <- NULL
+    if (includeAll == TRUE) {
+        result = subset(clProf.df, ID %in% result$ID)
+    }
+
+    ## remove zero count
+    result$Description <- as.character(result$Description) ## un-factor
+    GOlevel <- result[,c("ID", "Description")] ## GO ID and Term
+    GOlevel <- unique(GOlevel)
+
+    result <- result[result$Count != 0, ]
+    result$Description <- factor(result$Description,
+                                 levels=rev(GOlevel[,2]))
+
+
+    if (by=="rowPercentage") {
+        Description <- Count <- NULL # to satisfy codetools
+        result <- ddply(result,
+                        .(Description),
+                        transform,
+                        Percentage = Count/sum(Count),
+                        Total = sum(Count))
+
+        ## label GO Description with gene counts.
+        x <- mdply(result[, c("Description", "Total")], paste, sep=" (")
+        y <- sapply(x[,3], paste, ")", sep="")
+        result$Description <- y
+
+        ## restore the original order of GO Description
+        xx <- result[,c(2,3)]
+        xx <- unique(xx)
+        rownames(xx) <- xx[,1]
+        Termlevel <- xx[as.character(GOlevel[,1]),2]
+
+        ##drop the *Total* column
+        result <- result[, colnames(result) != "Total"]
+
+        result$Description <- factor(result$Description,
+                                     levels=rev(Termlevel))
+
+    } else if (by == "count") {
+        ## nothing
+    } else if (by == "geneRatio") {
+        gsize <- as.numeric(sub("/\\d+$", "", as.character(result$GeneRatio)))
+        gcsize <- as.numeric(sub("^\\d+/", "", as.character(result$GeneRatio)))
+        result$GeneRatio = gsize/gcsize
+        cluster <- paste(as.character(result$Cluster),"\n", "(", gcsize, ")", sep="")
+        lv <- unique(cluster)[order(as.numeric(unique(result$Cluster)))]
+        result$Cluster <- factor(cluster, levels = lv)
+    } else {
+        ## nothing
+    }
+    return(result)
+}
+
+
 ##' convert enrichResult object for ggplot2
 ##'
 ##'
 ##' @title fortify
-##' @param model enrichResult object
+##' @rdname fortify
+##' @param model 'enrichResult' or 'compareClusterResult' object
 ##' @param data not use here
 ##' @param showCategory Category numbers to show
 ##' @param by one of Count and GeneRatio
@@ -14,18 +129,21 @@
 ##' @importFrom ggplot2 fortify
 ##' @method fortify enrichResult
 ##' @export
-fortify.enrichResult <- function(model, data, showCategory=5, by = "Count", order=FALSE, drop=FALSE, split=NULL, ...) {
+fortify.enrichResult <- function(model, data, showCategory=5, by = "Count", order=FALSE,
+                                 drop=FALSE, split=NULL, ...) {
     fortify.internal(model, data, showCategory, by, order, drop, split, ...)
 }
 
 ##' @method fortify gseaResult
 ##' @export
-fortify.gseaResult <- function(model, data, showCategory=5, by = "Count", order=FALSE, drop=FALSE, split=NULL, ...) {
+fortify.gseaResult <- function(model, data, showCategory=5, by = "Count", order=FALSE,
+                               drop=FALSE, split=NULL, ...) {
     fortify.internal(model, data, showCategory, by, order, drop, split, ...)
 }
 
 
-fortify.internal <- function(model, data, showCategory=5, by = "Count", order=FALSE, drop=FALSE, split=NULL, ...) {
+fortify.internal <- function(model, data, showCategory=5, by = "Count", order=FALSE,
+                             drop=FALSE, split=NULL, ...) {
     res <- as.data.frame(model)
     res <- res[!is.na(res$Description), ]
     if (inherits(model, "gseaResult")) {